requirements.txt
neo4j
python-dotenv
PyPDF2
sentence_transformers
faiss-cpu
streamlit
langchain
python-dotenv
langchain_community
chromadb
pysqlite3-binary
pypdf
llama_index
llama-index-llms-ollama
llama-index-embeddings-huggingface

-U langchain-chroma
-U langchain-openai
